`Refactor ConfirmModal and Payment page to use InitialNotCloseConfirmModal`
